requirements.txt
django==1.10.4
django-debug-toolbar==1.4
djangorestframework==3.5.3

mongoengine==0.10.6
pymongo==3.3.1

django-rest-framework-mongoengine==3.3.0

# deployment
gunicorn==19.6.0
#aiohttp==1.1.5


git+https://github.com/zhwei820/dj-database-url.git
django-decouple

mysqlclient
djangorestframework-jwt
jsonschema